Fix crash if plugin using gtk3 is dlopen()d
If a library that is linked againset gtk3 (or just gobject, for that
matter) is dlopen()d, then RTLD_NEXT from a LD_PRELOADed library won't
work (unless that original program is also linked against gtk3).
Therefore, include a fallback that explicitly dlopen()s the library if
RTLD_NEXT doesn't work. This fixes a crash if this is used from
programs that use pggtk version 3.0.

Use RTLD_NOLOAD to make sure we don't spuriously load the library if
it wasn't there before. (Otherwise we'd mask linking problems with
plugins.)