Perform lookup of symbols we use at runtime
Some binaries may be compiled with ELF flag NOW (corresponding to
RTLD_NOW). If that is the case, symbol resolution is tried immediately
upon starting the binary by the dynamic linker. However, if the binary
itself is not linked against glib + gtk3, symbol lookup will fail.

Create stubs for symbols called from within the code that do the symbol
lookup at runtime, so that this library can always be loaded.

This fixes the issue of pulseaudio not starting if this library is
preloaded.