Implement CSD disablement for Gtk+ >= 3.16.1
This makes gtk3-nocsd work with Gtk+ >= 3.16.1. Beginning with that
version, the window manager hints to disable server side decorations
would always be set if a title bar is added to a window.

Since adding the title bar is a relatively simple task, we reimplement
it when overriding the set_titlebar function. The only tricky part is
that we need to figure out where to store the pointer to the title bar
object in the private data area of the GtkWindow - and since that
structure is private and not ABI-stable, we have to figure that out at
runtime.

This change is tested with various versions beginning with 3.14.5 up to
the current git development version of Gtk. It now successfully
disables CSD on all of them.