Also override GObject introspection called functions
Some functions that we need to override may also be called via GObject
introspection (e.g. when calling them from Python and other scripting
languages), so we need to make sure that also this method of calling
these functions causes our overrides to be called.

To this end, override the g_function_info_prep_invoker method to
replace the function pointers if necessary.