Try to load GDK functions we override from GDK2 if that's used.
In case of plugins (e.g. Python modules) that are dynamically loaded,
there already is code in place to dlopen() the proper libraries to make
sure that the the symbols are found (RTLD_NEXT won't work here).
However, gtk3-nocsd overrides some GDK functions that were already
available in GDK2 - and dlopen(GDK3) won't find anything if GDK2 is
used - but the LD_PRELOAD logic causes gtk3-nocsd's function (instead
of the original GDK2 one) to be called. This patch adds code to also
look for those two specific functions in GDK2 as well.

Fortunately, the two functions that are affected are binary compatible
between GDK2 and GDK3.

This fixes a crash mentioned in Pull Request #10 that wasn't fixed from
that pull request directly.