Wrapper: get rid of bashism, provide better error in some cases
In case the symlink to gtk3-nocsd is the only command present in PATH
with the correct name, instead of just falling back to /bin/false,
pretend to execute /.gtk3-nocsd./$APPNAME - so that the user will get
an error message that might help a little in figuring out what's wrong.

Also, use /bin/sh and not /bin/bash in shebang.