Use pthread_getspecific instead of __thread
When using LD_PRELOAD to load a library with __thread variables, a
slot in the DTV intended for static TLS is used, regardless of the
TLS model of the library that's preloaded. This is because a preloaded
library behaves like a linked library in this regard. Unfortunately,
the number of slots available for static TLS in the DTV is composed of
the number of slots required for libraries the binary was linked
against + a fixed number of overflow entries, the latter being
hard-coded into the dynamic linker. Without preloading anything, all of
the overflow slots will be free. The program is then able to load that
many plugins with a static TLS model via dlopen() successfully. If a
library with __thread is in LD_PRELOAD, it will use up an overflow slot,
such that one less plugin may be loaded by the program.

While dynamic plugins should never use a static TLS model to avoid
reaching the limit at all, there is some software that relies on the
fact that a limited number of slots is indeed available and when
loading certain plugins all are used up.

Therefore, in order not to break such programs (especially if
gtk3-nocsd is globally in LD_PRELOAD), do not use __thread for the
thread local storage, but rather resort to a traditional, purely
run-time implementation based on pthread_key_create and
pthread_getspecific.

Add a unit test that determines that there is no difference in the
number of libraries with static TLS enabled that may be preloaded by a
program, regardless of whether gtk3-nocsd is in LD_PRELOAD or not.

This fixes GitHub issue #11.