Selectively disable titlebar decorations
Instead of forcing GtkHeaderBar->show_close_button to FALSE, replace
the string used to add the standard decorations with a new string that
doesn't contain icon, minimize, maximize or close.

This means that the menu button will now be shown, if requested, and
not added as a single entry on an extra menubar, as has happened in
newer Gtk+3 versions (which makes the interface unintuitive, and in
older Gtk+3 versions the menu button was gone completely).

This means we have to override a lot more functions from GtkHeaderBar's
API though. Also, we need to hack g_object_get, as that is used from
within GtkHeaderBar to determine the default string from the global
GSettings when no specific one is set for a given widget.