Only cache gtk version check result if Gtk is loaded
If in a dynamical program (e.g. using python-gi) Glib is loaded before
Gtk, then the Gtk version check is executed before Gtk is even loaded,
returning FALSE and caching it. This will not disable CSD if Gtk is
loaded later. To circumvent this, cache the value only if we know that
Gtk is loaded.

Fixes github issue #16.