Properly handle the case when both gtk2 and gtk3 are loaded
In case both gtk2 and gtk3 are loaded, but only gtk2 is used (which can
happen if a piece of software is linked against neither, but has
different plugins, some of which link against gtk2, others against
gtk3, all loaded at the same time, but only one of the variants used),
make sure that gtk3-nocsd detects that, doesn't actually inject any
code and calls all the gtk2 functions, instead of the gtk3 variants.
(Because they are incompatible and mixing calls will lead to crashes.)

Fixes Github issue #18.